Refactor: Remove sync_csv.py and introduce data_fetcher.py, io_utils.py, ranking.py, and ratios.py
- Deleted sync_csv.py which handled Git synchronization and CSV file management.
- Added data_fetcher.py for fetching stock price data and financial information from Yahoo Finance.
- Introduced io_utils.py for utility functions to load and save YAML, CSV, and pickle files.
- Created ranking.py to rank financial ratios and aggregate scores.
- Added ratios.py to calculate various financial ratios based on raw data.
